// src/app/layout/vida-inversion/models/DataBeneficiary.ts
export class DataBeneficiary {

    sclient?: string;
    type_document?: any;
    document_number?: string;
    birthday_date?: any;
    names?: string;
    last_name?: string;
    last_name2?: string;
    gender?: any;
    civil_status?: any;
    nationality?: any;
    department?: any;
    province?: any;
    district?: any;
    phone?: string;
    email?: string;
    address?: string;
    assignment?: string;
    relation?: any;

    type_document_disabled?: boolean;
    document_number_disabled?: boolean;
    birthday_date_disabled?: boolean;
    names_disabled?: boolean;
    last_name_disabled?: boolean;
    last_name2_disabled?: boolean;
    gender_disabled?: boolean;
    civil_status_disabled?: boolean;
    nationality_disabled?: boolean;
    department_disabled?: boolean;
    province_disabled?: boolean;
    district_disabled?: boolean;
    phone_disabled?: boolean;
    email_disabled?: boolean;
    address_disabled?: boolean;
    relation_disabled?: boolean;
}